Fixed a bug that was caught by the ReallyBlocking test.
In occasions the thread that requested the operation would be the one
that ran it. This was compounded with the restriction that a coroutine
can only be resumed by the first thread that ran it means that it could
not be resumed by any other thread meant that the operation would become
defacto synchronous.

The fix consists adding a new field to the coroutine class that is used
to store what thread cannot run the coroutine. In this case, the thread
that requested the operation is the one that is forbidden from running
the operation coroutine.